+++ src/main/java/com/zxl/materialStorage/controller/storageManage/EsssController.java
@@ -1,10 +1,13 @@
 package com.zxl.materialStorage.controller.storageManage;
 
 import cn.hutool.core.util.ObjectUtil;
+import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.core.toolkit.StringUtils;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.zxl.materialStorage.common.api.ApiResult;
 import com.zxl.materialStorage.model.pojo.EssSpace;
+import com.zxl.materialStorage.model.pojo.MaterialEnter;
+import com.zxl.materialStorage.service.materialEnter.MaterialEnterService;
 import com.zxl.materialStorage.service.storageManage.EsssService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -24,6 +27,9 @@ import java.util.List;
 public class EsssController {
     @Autowired
     private EsssService esssService;
+
+    @Autowired
+    private MaterialEnterService materialEnterService;
 
     @PostMapping("/insertNewOne")
     public ApiResult<Object> insertNewOne(@RequestBody EssSpace essSpace){
@@ -45,6 +51,13 @@ public class EsssController {
             if (StringUtils.isEmpty(esssId)){
                 return ApiResult.blank();
             }
+            //若存在下级物资依赖，则拒绝删除
+            EssSpace byId = esssService.getById(esssId);
+            List<MaterialEnter> materialEnterList = materialEnterService.list(new QueryWrapper<MaterialEnter>().lambda().eq(MaterialEnter::getEsssNo, byId.getEsssNo()));
+            if (ObjectUtil.isNotEmpty(materialEnterList)){
+                return ApiResult.error();
+            }
+            //执行单个删除操作
             esssService.deleteOne(esssId);
         } catch (Exception e) {
             log.error("删除库区失败",e);
@@ -59,6 +72,15 @@ public class EsssController {
             if (ObjectUtil.isEmpty(esssIdList)){
                 return ApiResult.blank();
             }
+            //若存在下级物资依赖，则拒绝删除
+            List<EssSpace> essSpaceList = esssService.listByIds(esssIdList);
+            for (EssSpace space : essSpaceList) {
+                List<MaterialEnter> materialEnterList = materialEnterService.list(new QueryWrapper<MaterialEnter>().lambda().eq(MaterialEnter::getEsssNo, space.getEsssNo()));
+                if (ObjectUtil.isNotEmpty(materialEnterList)){
+                    return ApiResult.error();
+                }
+            }
+            //执行批量删除操作
             esssService.deleteMany(esssIdList);
         } catch (Exception e) {
             log.error("批量删除库区失败",e);
